requirements.txt
# Core AI and ML libraries
openai==1.12.0
anthropic==0.18.0
langchain==0.1.9
langchain-openai==0.0.8

# Web and API libraries
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
httpx==0.26.0

# Database and storage
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.27

# Data processing
pandas==2.2.0
numpy==1.26.4
pydantic==2.6.1

# Social media APIs
tweepy==4.14.0
python-twitter==3.5

# Utilities
python-dotenv==1.0.1
cryptography==42.0.2
pyjwt==2.8.0
click==8.1.7
rich==13.7.0

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.0.0

# Code quality
black==24.2.0
flake8==7.0.0
mypy==1.8.0

# Scheduling and background tasks
celery==5.3.6
schedule==1.2.1

# Image processing (optional)
pillow==10.2.0

